Split CovidTranslationPrApproval (which was doing two unrelated things) into two crons, the second is CovidDataAutoDeployment. They do two different things and run on different schedules.
