+++ CovidTranslationPrApproval/worker.js
